third_party/blink/web_tests/external/wpt/css/css-mixins/parameters-expected.txt

This is a testharness.js-based test.
[FAIL] Basic mixin parameter passing
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
[FAIL] Mixins can be called multiple times with different parameters
  assert_equals: expected "rgb(0, 0, 128)" but got "rgb(0, 0, 0)"
[FAIL] Mixin argument parsing with arguments wrapped in {}
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
[FAIL] Defaults in mixin parameters
  assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
